weather_project/requirements.txt
django-cors-headers
python-decouple
djangorestframework
requests
mongoengine